start mocking eodhd client
